Refactor and structure SEXP extensions.
