+++ lib/reek/ast/sexp_extensions/attribute_assignments.rb
@@ -1,0 +1,12 @@
+module Reek
+  module AST
+    module SexpExtensions
+      # Utility methods for :attrasgn nodes.
+      module AttrasgnNode
+        def args
+          children[2]
+        end
+      end
+    end
+  end
+end
