+++ lib/reek/ast/sexp_extensions/constant.rb
@@ -1,0 +1,12 @@
+module Reek
+  module AST
+    module SexpExtensions
+      # Utility methods for :const nodes.
+      module ConstNode
+        def simple_name
+          children.last
+        end
+      end
+    end
+  end
+end
