+++ lib/reek/ast/sexp_extensions/literal.rb
@@ -1,0 +1,12 @@
+module Reek
+  module AST
+    module SexpExtensions
+      # Utility methods for :lit nodes.
+      module LitNode
+        def value
+          children.first
+        end
+      end
+    end
+  end
+end
