+++ lib/reek/ast/sexp_extensions/super.rb
@@ -1,0 +1,14 @@
+module Reek
+  module AST
+    module SexpExtensions
+      # Utility methods for :super nodes.
+      module SuperNode
+        def method_name
+          :super
+        end
+      end
+
+      ZsuperNode = SuperNode
+    end
+  end
+end
