+++ lib/reek/ast/sexp_extensions/symbols.rb
@@ -1,0 +1,16 @@
+module Reek
+  module AST
+    module SexpExtensions
+      # Utility methods for :sym nodes.
+      module SymNode
+        def name
+          children.first
+        end
+
+        def full_name(outer)
+          "#{outer}##{name}"
+        end
+      end
+    end
+  end
+end
