+++ lib/reek/ast/sexp_extensions/when.rb
@@ -1,0 +1,16 @@
+module Reek
+  module AST
+    module SexpExtensions
+      # Utility methods for :when nodes.
+      module WhenNode
+        def condition_list
+          children[0..-2]
+        end
+
+        def body
+          children.last
+        end
+      end
+    end
+  end
+end
