+++ lib/reek/ast/sexp_extensions/yield.rb
@@ -1,0 +1,16 @@
+module Reek
+  module AST
+    module SexpExtensions
+      # Utility methods for :yield nodes.
+      module YieldNode
+        def args
+          children
+        end
+
+        def arg_names
+          args.map { |arg| arg[1] }
+        end
+      end
+    end
+  end
+end
